requirements.txt
absl-py==1.3.0
anyio @ file:///tmp/build/80754af9/anyio_1644481698350/work/dist
appdirs==1.4.4
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569684262/work
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
auditlog3==1.0.1
autopage==0.5.1
Babel @ file:///croot/babel_1671781930836/work
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.6.2
boltons @ file:///croot/boltons_1677628692245/work
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cliff==4.1.0
cmd2==2.4.3
comm==0.1.2
conda==23.3.1
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
debtcollector==2.5.0
debugpy==1.6.4
decorator==5.1.1
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Django==3.2.18
django-cors-headers==3.14.0
django-countries==7.5
django-filter==22.1
django-rapidapi==0.1.0.dev1
django-rest-knox==4.2.0
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dogpile.cache==1.1.8
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
filelock==3.12.0
Flask==2.3.1
flatbuffers==22.12.6
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.38.0
fsspec==2023.4.0
gast==0.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.51.1
h5py==2.10.0
heroku==0.1.4
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
iso8601==1.1.0
itsdangerous==2.1.2
jax==0.4.1
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==1.0.1
joblib==1.2.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonpatch==1.32
jsonpointer==2.3
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///croot/jupyter_console_1677674649642/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab @ file:///croot/jupyterlab_1675354114448/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1679055282532/work
jupyterlab_server @ file:///croot/jupyterlab_server_1677143054853/work
keras==2.11.0
keras-nightly==2.12.0.dev2022122308
Keras-Preprocessing==1.1.2
keystoneauth1==5.1.1
kiwisolver==1.4.4
koila==0.1.1
libclang==14.0.6
llvmlite==0.40.0rc1
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.4.2
mkdocs-material==5.5.14
mkdocs-material-extensions==1.1.1
msgpack==1.0.4
nbclassic @ file:///croot/nbclassic_1676902904845/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650308366712/work
nbconvert @ file:///croot/nbconvert_1668450669124/work
nbformat @ file:///croot/nbformat_1670352325207/work
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
networkx==3.0
nltk==3.8.1
notebook==6.0.3
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numba==0.47.0
numpy==1.21.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.68
openpyxl==3.0.10
openstacksdk==1.0.0
opt-einsum==3.3.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.1.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.1.0
packaging==22.0
pandas==1.5.3
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso==0.8.3
pbr==5.11.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
platformdirs==2.6.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648042571233/work
ply==3.11
prettytable==3.6.0
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pymdown-extensions==9.11
pynvml==11.5.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5-sip==12.11.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110947380/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-cinderclient==9.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-decouple==3.7
python-keystoneclient==5.0.1
python-novaclient==18.2.0
python-openstackclient==6.0.0
pytz==2022.7
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
qtconsole @ file:///croot/qtconsole_1674008428467/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requestsexceptions==1.4.0
rfc3986==2.0.0
rich==13.3.4
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scikit-learn==1.2.0
scipy==1.4.1
seaborn==0.12.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sentry-sdk==1.19.1
simplejson==3.18.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio @ file:///tmp/build/80754af9/sniffio_1614030475067/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
sqlparse==0.4.3
stack-data==0.6.2
stevedore==4.1.1
tb-nightly==2.12.0a20221222
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado @ file:///croot/terminado_1671751832461/work
tf-estimator-nightly==2.12.0.dev2022122209
tf-nightly==2.12.0.dev20221222
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2 @ file:///croot/tinycss2_1668168815555/work
tokenizers==0.13.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
toolz @ file:///croot/toolz_1667464077321/work
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers==4.28.1
typing_extensions==4.4.0
ultralytics==8.0.77
urllib3==1.26.14
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1614804261064/work
Werkzeug==2.3.2
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
wrapt==1.14.1
zipp==3.11.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
